New html doc building works

+++ test/SqlScripts/sql/Address.sql
@@ -1,5 +1,5 @@
 CREATE TABLE [Address] (
-  [aId] INT, 
+  [aId] BIGINT, 
   [aStreet1] VARCHAR(50), 
   [aStreet2] VARCHAR(50), 
   [aCity] VARCHAR(50),
